Removing ItemDefinition ID objects. Instead definitions  will be passed around at runtime. This reduces the complexity of having to manage a type for an id and a definition.
This does introduce possible risk of putting the wrong items into slots that they should not be in such as a resource going into a drill slot on a miner.
I will need to come up with another way of handling this. It might just be... don't write stupid game configs since I will be manually writing them. Maybe add a validator somewhere.